+++ android/src/main/java/top/kikt/ijkplayer/NotifyChannel.kt
@@ -32,7 +32,8 @@ class NotifyChannel(val registry: PluginRegistry.Registrar, val textureId: Long,
         }
         player.setOnBufferingUpdateListener { mp, percent ->
             /// 在线视频缓冲
-            logi("completion buffer update $info $percent")
+         //   logi("completion buffer update $info $percent")
+           channel.invokeMethod("setOnBufferingUpdate", info)
         }
         player.setOnSeekCompleteListener {
             logi("onSeekCompletion $info")