+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-## 0.1.2
+## 0.1.4
+
+IjkMediaController  添加新属性  isNomal  //是否是单个视频（用于区分视频与列表默认是true）
+主要解决在视频列表中要禁止上下手势滑动
+在example的自定义布局中添加新属性：
+
+DefaultIJKControllerWidget
+ //新添加控制器属性   广告用
+  final String adimageUrl; //广告图片
+  final String adTitle; //广告标题
+  final double adrevealTime; //显示时间
+  final double addisappearTime; //消失时间
+  final bool isShowAD; //是否显示广告  true  //显示  false 隐藏
+  final bool isShowRatio; //是否显示分辨率的选项  true  //显示  false 隐藏
+  final String videoTitleTxT; //视频标题 
+## 0.1.3
 
 fix some bug
 ## 0.1.2